:recycle: 웰컴딜 API 컴포넌트로 이동
